+++ tests/integration/components/tooltipper/reference-element-test.js
@@ -7,7 +7,7 @@ module('tooltipper', function(hooks) {
   setupTooltipperTest(hooks);
 
   test('reference element', async function(assert) {
-    assert.expect(2);
+    assert.expect(4);
 
     await render(hbs`
       {{#if this.referenceElement}}
@@ -16,18 +16,19 @@ module('tooltipper', function(hooks) {
           @referenceElement={{this.referenceElement}} />
       {{/if}}
 
-      <div class="reference-element"></div>
+      <div class="reference-element-1"></div>
+      <div class="reference-element-2"></div>
     `);
 
-    this.set('referenceElement', find('.reference-element'));
+    this.set('referenceElement', find('.reference-element-1'));
 
-    await triggerEvent('.reference-element', 'mouseenter');
+    await triggerEvent('.reference-element-1', 'mouseenter');
 
     assert
       .dom('.tooltip')
       .exists('renders tooltip when mousing over the reference element');
 
-    await triggerEvent('.reference-element', 'mouseleave');
+    await triggerEvent('.reference-element-1', 'mouseleave');
 
     await triggerEvent('.tooltip', 'animationend');
 
@@ -36,5 +37,21 @@ module('tooltipper', function(hooks) {
       .doesNotExist(
         'tooltip is destroyed when mousing out of the reference element'
       );
+
+    this.set('referenceElement', find('.reference-element-2'));
+
+    await triggerEvent('.reference-element-1', 'mouseenter');
+
+    assert
+      .dom('.tooltip')
+      .doesNotExist(
+        'tooltip is not displayed when hovering over original reference element'
+      );
+
+    await triggerEvent('.reference-element-2', 'mouseenter');
+
+    assert
+      .dom('.tooltip')
+      .exists('renders tooltip when mousing over the new reference element');
   });
 });